# Improving counting performance