--- a/git-pack/src/cache.rs
+++ b/git-pack/src/cache.rs
@@ -104,6 +104,91 @@ pub mod lru {
     #[cfg(feature = "pack-cache-lru-dynamic")]
     pub use memory::MemoryCappedHashmap;
 
+    #[cfg(feature = "pack-cache-lru-dynamic")]
+    mod memory2 {
+        use super::DecodeEntry;
+        use clru::WeightScale;
+        use std::num::NonZeroUsize;
+
+        struct Entry {
+            data: Vec<u8>,
+            kind: git_object::Kind,
+            compressed_size: usize,
+        }
+
+        type Key = (u32, u64);
+        struct CustomScale;
+
+        impl WeightScale<Key, Entry> for CustomScale {
+            fn weight(&self, _key: &Key, value: &Entry) -> usize {
+                value.data.len()
+            }
+        }
+
+        /// An LRU cache with hash map backing and an eviction rule based on the memory usage for object data in bytes.
+        pub struct MemoryCappedHashmap2 {
+            inner: clru::CLruCache<Key, Entry, std::collections::hash_map::RandomState, CustomScale>,
+            free_list: Vec<Vec<u8>>,
+            debug: git_features::cache::Debug,
+        }
+
+        impl MemoryCappedHashmap2 {
+            /// Return a new instance which evicts least recently used items if it uses more than `memory_cap_in_bytes`
+            /// object data.
+            pub fn new(memory_cap_in_bytes: usize) -> MemoryCappedHashmap2 {
+                MemoryCappedHashmap2 {
+                    inner: clru::CLruCache::with_config(
+                        clru::CLruCacheConfig::new(NonZeroUsize::new(memory_cap_in_bytes).expect("non zero"))
+                            .with_scale(CustomScale),
+                    ),
+                    free_list: Vec::new(),
+                    debug: git_features::cache::Debug::new(format!("MemoryCappedHashmap({}B)", memory_cap_in_bytes)),
+                }
+            }
+        }
+
+        impl DecodeEntry for MemoryCappedHashmap2 {
+            fn put(&mut self, pack_id: u32, offset: u64, data: &[u8], kind: git_object::Kind, compressed_size: usize) {
+                self.debug.put();
+                if let Ok(Some(previous_entry)) = self.inner.put_with_weight(
+                    (pack_id, offset),
+                    Entry {
+                        data: self
+                            .free_list
+                            .pop()
+                            .map(|mut v| {
+                                v.clear();
+                                v.resize(data.len(), 0);
+                                v.copy_from_slice(data);
+                                v
+                            })
+                            .unwrap_or_else(|| Vec::from(data)),
+                        kind,
+                        compressed_size,
+                    },
+                ) {
+                    self.free_list.push(previous_entry.data)
+                }
+            }
+
+            fn get(&mut self, pack_id: u32, offset: u64, out: &mut Vec<u8>) -> Option<(git_object::Kind, usize)> {
+                let res = self.inner.get(&(pack_id, offset)).map(|e| {
+                    out.resize(e.data.len(), 0);
+                    out.copy_from_slice(&e.data);
+                    (e.kind, e.compressed_size)
+                });
+                if res.is_some() {
+                    self.debug.hit()
+                } else {
+                    self.debug.miss()
+                }
+                res
+            }
+        }
+    }
+    #[cfg(feature = "pack-cache-lru-dynamic")]
+    pub use memory2::MemoryCappedHashmap2;
+
     #[cfg(feature = "pack-cache-lru-static")]
     mod _static {
         use super::DecodeEntry;