--- a/gitoxide-core/src/pack/create.rs
+++ b/gitoxide-core/src/pack/create.rs
@@ -210,6 +210,7 @@ where
         counts.shrink_to_fit();
         counts
     };
+    todo!("figure out performance issues");
 
     progress.inc();
     let num_objects = counts.len();