… Cargo.lock
… git-odb/Cargo.toml
… git-pack/Cargo.toml
… git-pack/tests/pack/data/output/count_and_entries.rs
… git-repository/Cargo.toml